fix(frontend): return 404 status codes when route is not found
